+++ src/mantle/cache/class-cache-service-provider.php
@@ -17,9 +17,6 @@ class Cache_Service_Provider extends Service_Provider {
 	 * Register the service provider.
 	 */
 	public function register(): void {
-		$this->app->singleton(
-			'cache',
-			fn ( $app) => new Cache_Manager( $app )
-		);
+		$this->app->singleton( 'cache', fn ( $app ) => new WordPress_Cache_Repository( $app ) );
 	}
 }