+++ src/mantle/cache/class-repository.php
@@ -10,15 +10,14 @@
 namespace Mantle\Cache;
 
 use Closure;
-
-use function Mantle\Support\Helpers\collect;
+use Psr\SimpleCache\CacheInterface;
 
 /**
  * Cache repository.
  *
  * This class contains some camelCase methods to match the PSR interface.
  */
-abstract class Repository {
+abstract class Repository implements CacheInterface {
 	/**
 	 * Retrieve a value from cache.
 	 *
@@ -31,29 +30,79 @@ abstract class Repository {
 	abstract public function get( string $key, mixed $default = null ): mixed;
 
 	/**
-	 * Remove an item from the cache.
+	 * Retrieve multiple cache keys.
 	 *
-	 * @param  string $key Cache key.
+	 * @param iterable $keys Cache keys.
+	 * @param mixed    $default Default value.
 	 */
-	abstract public function forget( $key ): bool;
+	abstract public function get_multiple( iterable $keys, mixed $default = null ): iterable;
 
 	/**
-	 * Store an item in the cache.
+	 * Retrieve a value from cache. Return it if it exists and if stale, refresh it after the response is sent.
 	 *
-	 * @param  string                                    $key
-	 * @param  mixed                                     $value
-	 * @param  \DateTimeInterface|\DateInterval|int|null $ttl
+	 * @template TCacheValue
+	 *
+	 * @param string                                    $key
+	 * @param int|\DateInterval|\DateTimeInterface|null $stale
+	 * @param int|\DateInterval|\DateTimeInterface|null $expire
+	 * @param Closure(): TCacheValue                    $callback
+	 * @return (TCacheValue is null ? mixed : TCacheValue)
 	 */
-	abstract public function put( $key, $value, $ttl = null ): bool;
+	abstract public function flexible( string $key, int|\DateInterval|\DateTimeInterface|null $stale, int|\DateInterval|\DateTimeInterface|null $expire, callable $callback ): mixed;
+
+	/**
+	 * Set a cache item.
+	 *
+	 * @param string                 $key Cache key.
+	 * @param mixed                  $value Item value.
+	 * @param null|int|\DateInterval $ttl TTL.
+	 */
+	abstract public function set( string $key, mixed $value, int|\DateInterval|null $ttl = null ): bool;
+
+	/**
+	 * Set multiple keys.
+	 *
+	 * @param iterable               $values Key value pair of values to set.
+	 * @param null|int|\DateInterval $ttl Cache TTL.
+	 */
+	abstract public function set_multiple( iterable $values, null|int|\DateInterval $ttl = null ): bool;
+
+	/**
+	 * Delete a cache key.
+	 *
+	 * @param string $key Cache key.
+	 */
+	abstract public function delete( string $key ): bool;
+
+	/**
+	 * Delete multiple cache keys.
+	 *
+	 * @param iterable<string> $keys Cache keys.
+	 */
+	abstract public function delete_multiple( iterable $keys ): bool;
+
+	/**
+	 * Alias for flexible.
+	 *
+	 * @template TCacheValue
+	 *
+	 * @param string                                    $key
+	 * @param int|\DateInterval|\DateTimeInterface|null $stale
+	 * @param int|\DateInterval|\DateTimeInterface|null $expire
+	 * @param Closure(): TCacheValue                    $callback
+	 * @return (TCacheValue is null ? mixed : TCacheValue)
+	 */
+	public function swr( string $key, int|\DateInterval|\DateTimeInterface|null $stale, int|\DateInterval|\DateTimeInterface|null $expire, Closure $callback ): mixed {
+		return $this->flexible( $key, $stale, $expire, $callback );
+	}
 
 	/**
 	 * Retrieve an item from the cache and delete it.
 	 *
 	 * @param  string $key
 	 * @param  mixed  $default
-	 * @return mixed
 	 */
-	public function pull( $key, $default = null ) {
+	public function pull( string $key, mixed $default = null ): mixed {
 		$value = $this->get( $key );
 
 		if ( ! is_null( $value ) ) {
@@ -66,14 +115,42 @@ abstract class Repository {
 	}
 
 	/**
+	 * Alias for get_multiple for the PSR interface.
+	 *
+	 * @param iterable $keys Cache keys.
+	 * @param mixed    $default Default value.
+	 */
+	public function getMultiple( iterable $keys, mixed $default = null ): iterable {
+		return $this->get_multiple( $keys, $default );
+	}
+
+	/**
+	 * Alias for set_multiple for the PSR interface.
+	 *
+	 * @param iterable               $values Key value pair of values to set.
+	 * @param null|int|\DateInterval $ttl Cache TTL.
+	 */
+	public function setMultiple( iterable $values, null|int|\DateInterval $ttl = null ): bool {
+		return $this->set_multiple( $values, $ttl );
+	}
+
+	/**
+	 * Alias for delete_multiple for the PSR interface.
+	 *
+	 * @param iterable<string> $keys Cache keys.
+	 */
+	public function deleteMultiple( iterable $keys ): bool {
+		return $this->delete_multiple( $keys );
+	}
+
+	/**
 	 * Store an item in the cache if the key does not exist.
 	 *
 	 * @param  string                                    $key
 	 * @param  mixed                                     $value
 	 * @param  \DateTimeInterface|\DateInterval|int|null $ttl
-	 * @return bool
 	 */
-	public function add( $key, $value, $ttl = null ) {
+	public function add( string $key, mixed $value, \DateTimeInterface|\DateInterval|int|null $ttl = null ): bool {
 		if ( is_null( $this->get( $key ) ) ) {
 			return $this->put( $key, $value, $ttl );
 		}
@@ -82,64 +159,23 @@ abstract class Repository {
 	}
 
 	/**
-	 * Set a cache item.
+	 * Alias for set.
 	 *
-	 * @param string                 $key Cache key.
-	 * @param mixed                  $value Item value.
-	 * @param null|int|\DateInterval $ttl TTL.
+	 * @param string                                    $key Cache key.
+	 * @param mixed                                     $value Item value.
+	 * @param \DateTimeInterface|\DateInterval|int|null $ttl Cache TTL.
 	 */
-	public function set( string $key, mixed $value, null|int|\DateInterval $ttl = null ): bool {
-		return $this->put( $key, $value, $ttl );
+	public function put( string $key, mixed $value, \DateTimeInterface|\DateInterval|int|null $ttl = null ): bool {
+		return $this->set( $key, $value, $ttl );
 	}
 
 	/**
-	 * Delete a cache key.
+	 * Alias for delete.
 	 *
 	 * @param string $key Cache key.
 	 */
-	public function delete( string $key ): bool {
-		return $this->forget( $key );
-	}
-
-	/**
-	 * Retrieve multiple cache keys.
-	 *
-	 * @param iterable $keys Cache keys.
-	 * @param mixed    $default Default value.
-	 */
-	public function getMultiple( iterable $keys, mixed $default = null ): iterable {
-		return collect( $keys )
-			->map(
-				fn ( $key) => $this->pull( $key, $default )
-			)
-			->to_array();
-	}
-
-	/**
-	 * Set multiple keys.
-	 *
-	 * @param iterable               $values Key value pair of values to set.
-	 * @param null|int|\DateInterval $ttl Cache TTL.
-	 */
-	public function setMultiple( iterable $values, null|int|\DateInterval $ttl = null ): bool {
-		foreach ( $values as $key => $value ) {
-			$this->set( $key, $value, $ttl );
-		}
-
-		return true;
-	}
-
-	/**
-	 * Delete multiple cache keys.
-	 *
-	 * @param iterable<string> $keys Cache keys.
-	 */
-	public function deleteMultiple( iterable $keys ): bool {
-		collect( $keys )->each(
-			fn ( $key) => $this->delete( $key )
-		);
-
-		return true;
+	public function forget( string $key ): bool {
+		return $this->delete( $key );
 	}
 
 	/**
@@ -156,9 +192,8 @@ abstract class Repository {
 	 *
 	 * @param  string $key Cache key.
 	 * @param  mixed  $value Value.
-	 * @return bool
 	 */
-	public function forever( $key, $value ) {
+	public function forever( string $key, mixed $value ): bool {
 		return $this->put( $key, $value, null );
 	}
 
@@ -168,9 +203,8 @@ abstract class Repository {
 	 * @param  string                                    $key
 	 * @param  \DateTimeInterface|\DateInterval|int|null $ttl
 	 * @param  \Closure                                  $callback
-	 * @return mixed
 	 */
-	public function remember( $key, $ttl, Closure $callback ) {
+	public function remember( string $key, \DateTimeInterface|\DateInterval|int|null $ttl, Closure $callback ): mixed {
 		$value = $this->get( $key );
 
 		if ( ! is_null( $value ) ) {
@@ -185,20 +219,8 @@ abstract class Repository {
 	 *
 	 * @param  string   $key Cache key.
 	 * @param  \Closure $callback Callback to invoke.
-	 * @return mixed
 	 */
-	public function sear( $key, Closure $callback ) {
-		return $this->rememberForever( $key, $callback );
-	}
-
-	/**
-	 * Get an item from the cache, or execute the given Closure and store the result forever.
-	 *
-	 * @param  string   $key Cache key.
-	 * @param  \Closure $callback Callback to invoke.
-	 * @return mixed
-	 */
-	public function rememberForever( $key, Closure $callback ) {
+	public function remember_forever( string $key, Closure $callback ): mixed {
 		$value = $this->get( $key );
 
 		if ( ! is_null( $value ) ) {
@@ -206,5 +228,15 @@ abstract class Repository {
 		}
 
 		return $this->put( $key, $callback(), null );
+	}
+
+	/**
+	 * Get an item from the cache, or execute the given Closure and store the result forever.
+	 *
+	 * @param  string   $key Cache key.
+	 * @param  \Closure $callback Callback to invoke.
+	 */
+	public function sear( string $key, Closure $callback ): mixed {
+		return $this->remember_forever( $key, $callback );
 	}
 }